Cleanup for pull request
